// types.ts
export interface ExperienceItem {
  id: string;
  role: string;
  company: string;
  timeRange: string;
  logoUrl: string;
}

export interface EducationItem {
  id: string;
  degree: string;
  school: string;
  timeRange: string;
  logoUrl: string;
}

export interface ProjectItem {
  id: string;
  name: string;
  description: string;
  imageUrl: string;
}